+++ Clockwork/DataSource/EloquentDataSource.php
@@ -1,10 +1,8 @@
 <?php namespace Clockwork\DataSource;
 
-use Clockwork\Helpers\Serializer;
-use Clockwork\Helpers\StackTrace;
+use Clockwork\Helpers\{Serializer, StackTrace};
 use Clockwork\Request\Request;
-use Clockwork\Support\Laravel\Eloquent\ResolveModelLegacyScope;
-use Clockwork\Support\Laravel\Eloquent\ResolveModelScope;
+use Clockwork\Support\Laravel\Eloquent\{ResolveModelLegacyScope, ResolveModelScope};
 
 use Illuminate\Database\ConnectionResolverInterface;
 use Illuminate\Contracts\Events\Dispatcher as EventDispatcher;