+++ Clockwork/Request/IncomingRequest.php
@@ -8,6 +8,9 @@ class IncomingRequest
 	// URI
 	public $uri;
 
+	// Headers
+	public $headers;
+
 	// GET and POST data
 	public $input = [];
 	// Cookies
@@ -19,6 +22,18 @@ class IncomingRequest
 	public function __construct(array $data = [])
 	{
 		foreach ($data as $key => $val) $this->$key = $val;
+	}
+
+	// Returns a header value, or default if not set
+	public function header($key, $default = null)
+	{
+		return $this->headers[$key] ?? $default;
+	}
+
+	// Returns an input value, or default if not set
+	public function input($key, $default = null)
+	{
+		return $this->input[$key] ?? $default;
 	}
 
 	// Returns true, if HTTP host is one of the common domains used for local development