+++ Clockwork/Storage/Search.php
@@ -1,7 +1,6 @@
 <?php namespace Clockwork\Storage;
 
-use Clockwork\Request\Request;
-use Clockwork\Request\RequestType;
+use Clockwork\Request\{Request, RequestType};
 
 // Rules for searching requests
 class Search
@@ -23,11 +22,11 @@ class Search
 	public function __construct($search = [], $options = [])
 	{
 		foreach ([ 'uri', 'controller', 'method', 'status', 'time', 'received', 'name', 'type' ] as $condition) {
-			$this->$condition = isset($search[$condition]) ? $search[$condition] : [];
+			$this->$condition = $search[$condition] ?? [];
 		}
 
 		foreach ([ 'stopOnFirstMismatch' ] as $option) {
-			$this->$option = isset($options[$option]) ? $options[$option] : $this->$option;
+			$this->$option = $options[$option] ?? $this->$option;
 		}
 
 		$this->method = array_map('strtolower', $this->method);