+++ Clockwork/Support/Laravel/Console/CapturingFormatter.php
@@ -1,9 +1,8 @@
 <?php namespace Clockwork\Support\Laravel\Console;
 
-use Symfony\Component\Console\Formatter\OutputFormatterInterface;
-use Symfony\Component\Console\Formatter\OutputFormatterStyleInterface;
+use Symfony\Component\Console\Formatter\{OutputFormatterInterface, OutputFormatterStyleInterface};
 
-// Formatter wrapping around a "real" formatter, capturing the formatted output (Symfony 6.x and later)
+// Formatter wrapping around a "real" formatter, capturing the formatted output (Symfony 7.x and later)
 class CapturingFormatter implements OutputFormatterInterface
 {
 	protected $formatter;
@@ -24,9 +23,9 @@ class CapturingFormatter implements OutputFormatterInterface
 		return $capturedOutput;
 	}
 
-	public function setDecorated(bool $decorated)
+	public function setDecorated(bool $decorated): void
 	{
-		return $this->formatter->setDecorated($decorated);
+		$this->formatter->setDecorated($decorated);
 	}
 
 	public function isDecorated(): bool
@@ -34,9 +33,9 @@ class CapturingFormatter implements OutputFormatterInterface
 		return $this->formatter->isDecorated();
 	}
 
-	public function setStyle(string $name, OutputFormatterStyleInterface $style)
+	public function setStyle(string $name, OutputFormatterStyleInterface $style): void
 	{
-		return $this->formatter->setStyle($name, $style);
+		$this->formatter->setStyle($name, $style);
 	}
 
 	public function hasStyle(string $name): bool