+++ Clockwork/Support/Laravel/config/clockwork.php
@@ -72,6 +72,17 @@ return [
 			],
 		],
 
+		// Sent HTTP requests
+		'http_requests' => [
+			'enabled' => env('CLOCKWORK_HTTP_REQUESTS_ENABLED', true),
+
+			// Collect structured request and response content (json and form data)
+			'collect_data' => env('CLOCKWORK_HTTP_REQUESTS_COLLECT_DATA', true),
+
+			// Collect raw request and response content (high storage usage with large responses)
+			'collect_raw_data' => env('CLOCKWORK_HTTP_REQUESTS_COLLECT_RAW_DATA', false)
+		],
+
 		// Laravel log (you can still log directly to Clockwork with laravel log disabled)
 		'log' => [
 			'enabled' => env('CLOCKWORK_LOG_ENABLED', true)