+++ Clockwork/Support/Lumen/Controller.php
@@ -3,9 +3,7 @@
 use Clockwork\Clockwork;
 use Clockwork\Support\Lumen\ClockworkSupport;
 
-use Illuminate\Http\JsonResponse;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
+use Illuminate\Http\{JsonResponse, RedirectResponse, Request};
 use Laravel\Lumen\Routing\Controller as LumenController;
 use Laravel\Telescope\Telescope;