+++ Clockwork/Support/Symfony/ClockworkSupport.php
@@ -1,14 +1,11 @@
 <?php namespace Clockwork\Support\Symfony;
 
-use Clockwork\Authentication\NullAuthenticator;
-use Clockwork\Authentication\SimpleAuthenticator;
+use Clockwork\Authentication\{NullAuthenticator, SimpleAuthenticator};
 use Clockwork\Storage\Search;
 use Clockwork\Web\Web;
 
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\HttpFoundation\BinaryFileResponse;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\{BinaryFileResponse, JsonResponse, Request};
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class ClockworkSupport